move steps inside correct folders
