+++ commands/create_symfony_project_command.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/racastellanosm/tools-php-skeleton-creator/utilities"
 	"github.com/racastellanosm/tools-php-skeleton-creator/workflows"
+	"github.com/racastellanosm/tools-php-skeleton-creator/workflows/steps"
 	"github.com/spf13/cobra"
 )
 
@@ -16,10 +17,9 @@ var (
 var CreateProjectWithSymfonyCommand = &cobra.Command{
 	Use:   "create-with-symfony --name your-project-name [--db postgresql|mysql ]",
 	Short: "Create a new PHP project with Symfony Framework",
-	Long: `Create a new PHP project with Symfony framework following recommended development guidelines (DDD, CQRS, Testing, Automation).
-This command initializes a new project directory with the necessary files and structure for a PHP application.`,
+	Long:  `Create a new PHP project with Symfony framework following recommended development guidelines (DDD, CQRS, Testing, Automation).\nThis command initializes a new project directory with the necessary files and structure for a PHP application.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		workflowParameters := workflows.StepParameters{
+		workflowParameters := steps.StepParameters{
 			ProjectName: projectName,
 			Database:    database,
 		}
