+++ workflows/steps/slim/reorganize_needed_folders_step.go
@@ -1,4 +1,4 @@
-package create
+package slim
 
 import (
 	"fmt"
