+++ workflows/steps/slim/require_slim_skeleton_step.go
@@ -1,4 +1,4 @@
-package create
+package slim
 
 import (
 	"fmt"
