+++ workflows/steps/symfony/reorganize_needed_folders_step.go
@@ -1,4 +1,4 @@
-package create
+package symfony
 
 import (
 	"fmt"
