+++ workflows/steps/symfony/require_additional_dependencies_step.go
@@ -1,4 +1,4 @@
-package create
+package symfony
 
 import (
 	"fmt"
