+++ workflows/steps/symfony/require_symfony_skeleton_step.go
@@ -1,4 +1,4 @@
-package create
+package symfony
 
 import (
 	"fmt"
