+++ workflows/steps/symfony/rewrite_yaml_config_to_php_step.go
@@ -1,4 +1,4 @@
-package create
+package symfony
 
 import (
 	"fmt"
